Raxol.Terminal.Parser.State.Manager (Raxol Terminal v2.4.0)

Manages the state of the terminal parser, including escape sequences, control sequences, and parser modes.

Types

intermediate()

@type intermediate() :: [non_neg_integer()]

params()

@type params() :: [non_neg_integer()]

parser_state()

@type parser_state() ::
  :ground
  | :escape
  | :csi_entry
  | :csi_param
  | :csi_intermediate
  | :csi_ignore
  | :osc_string
  | :dcs_entry
  | :dcs_param
  | :dcs_intermediate
  | :dcs_passthrough
  | :apc_string
  | :pm_string
  | :sos_string
  | :string

string_flags()

@type string_flags() :: %{required(String.t()) => boolean()}

t()

@type t() :: %Raxol.Terminal.Parser.State.Manager{
  apc_buffer: String.t(),
  dcs_buffer: String.t(),
  designating_gset: term(),
  final_byte: term(),
  ignore: boolean(),
  intermediate: intermediate(),
  intermediates_buffer: term(),
  osc_buffer: String.t(),
  params: params(),
  params_buffer: term(),
  payload_buffer: term(),
  pm_buffer: String.t(),
  single_shift: term(),
  sos_buffer: String.t(),
  state: parser_state(),
  string_buffer: String.t(),
  string_flags: string_flags(),
  string_parser_state: parser_state() | nil,
  string_terminator: non_neg_integer() | nil
}

Functions

append_intermediate(manager, intermediate)

append_param(manager, param)

append_payload(manager, payload)

clear_string_buffers(manager)

Clears all string buffers.

get_apc_buffer(manager)

Gets the APC buffer content.

get_current_state(manager)

get_dcs_buffer(manager)

Gets the DCS buffer content.

get_intermediate(manager)

Gets the current intermediate characters.

get_osc_buffer(manager)

Gets the OSC buffer content.

get_params(manager)

Gets the current parameters.

get_pm_buffer(manager)

Gets the PM buffer content.

get_sos_buffer(manager)

Gets the SOS buffer content.

get_state(manager)

Gets the current parser state.

get_string_buffer(manager)

Gets the string buffer content.

get_string_flags(manager)

Gets the string flags.

get_string_parser_state(manager)

Gets the string parser state.

get_string_terminator(manager)

Gets the string terminator.

ignore?(manager)

Checks if the parser is in ignore mode.

new()

Creates a new parser state manager instance.

process_char(manager, char)

Processes a single character and updates the parser state accordingly.

process_input(emulator, state, input)

reset(manager)

Resets the parser state manager to its initial state.

set_apc_buffer(manager, content)

Sets the APC buffer content.

set_dcs_buffer(manager, content)

Sets the DCS buffer content.

set_designating_gset(manager, gset)

set_final_byte(manager, byte)

set_ignore(manager, ignore)

Sets the ignore mode.

set_intermediate(manager, intermediate)

Sets the intermediate characters.

set_osc_buffer(manager, content)

Sets the OSC buffer content.

set_params(manager, params)

Sets the parameters.

set_pm_buffer(manager, content)

Sets the PM buffer content.

set_sos_buffer(manager, content)

Sets the SOS buffer content.

set_state(manager, state)

Sets the parser state.

set_string_buffer(manager, content)

Sets the string buffer content.

set_string_flags(manager, flags)

Sets the string flags.

set_string_parser_state(manager, state)

Sets the string parser state.

set_string_terminator(manager, terminator)

Sets the string terminator.

transition_to(manager, arg2)